+++ django/jwt/utils.py
@@ -780,10 +780,7 @@ def get_admin_logout_urlpattern(
                 f"{get_script_prefix()}{auth_service_prefix}/admin/login/"
             )
 
-            cookie_name = getattr(settings, "JWT_COOKIE_NAME", "stapel_jwt")
-            refresh_cookie_name = getattr(
-                settings, "JWT_REFRESH_COOKIE_NAME", "stapel_refresh_jwt"
-            )
+            cookie_name, refresh_cookie_name = jwt_cookie_names()
             cookie_domain = getattr(settings, "JWT_COOKIE_DOMAIN", None)
             cookie_samesite = getattr(settings, "JWT_COOKIE_SAMESITE", "Lax")